+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 ## [Unreleased]
 
 - Update `protoc@latest` to 3.21.12.
+- Support aarch64 self-hosted runners (Linux, macOS, Windows).
 - Improve support for Fedora/RHEL based containers/self-hosted runners.
 
 ## [1.16.0] - 2022-12-14